+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-06-04 12:19:56
-// Commit: feat(security): add bi-directional event bus handler for security
+// Updated at: 2026-06-04 22:08:25
+// Commit: fix(security): prevent double tap duplicate action in security
 
 class securityComponentState {
-  final String commitRef = '1744';
-  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
-  final DateTime timestamp = DateTime.parse('2026-06-04T12:19:56+05:00');
+  final String commitRef = '1751';
+  final String lastMessage = 'fix(security): prevent double tap duplicate action in security';
+  final DateTime timestamp = DateTime.parse('2026-06-04T22:08:25+05:00');
 }